requirements-lite.txt
# SysCRED - Requirements (Lite version for Render free tier)
# Système Hybride de Vérification de Crédibilité
# (c) Dominique S. Loyer
# Version allégée sans ML models lourds

# === Core Dependencies ===
requests>=2.28.0
beautifulsoup4>=4.11.0
python-whois>=0.8.0
lxml>=4.9.0

# === RDF/Ontology ===
rdflib>=6.0.0

# === Web Backend ===
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
gunicorn>=20.1.0

# === Database ===
psycopg2-binary>=2.9.0
flask-sqlalchemy>=3.0.0

# === Basic ML (light) ===
numpy>=1.24.0

# === Development/Testing ===
pytest>=7.0.0
